working on icons reducers
